+++ source/Nuke.GlobalTool/Program.AddPackage.cs
@@ -23,7 +23,8 @@ namespace Nuke.GlobalTool
         {
             var packageId = args.ElementAt(0);
             var packageVersion =
-                (args.ElementAtOrDefault(1) ??
+                (EnvironmentInfo.GetParameter<string>("version") ??
+                 args.ElementAtOrDefault(1) ??
                  NuGetPackageResolver.GetLatestPackageVersion(packageId, includePrereleases: false).GetAwaiter().GetResult() ??
                  NuGetPackageResolver.GetGlobalInstalledPackage(packageId, version: null, packagesConfigFile: null)?.Version.ToString())
                 .NotNull("packageVersion != null");